refactor: Separate module for music.
