+++ src/sfx/sfx.ts
@@ -36,12 +36,18 @@ import {
     // @ts-expect-error JS code exported by a tool
 } from "./sfxData.js";
 
-import { isIPad, isDesktop, isIOS } from "../core/platform/deviceDetection";
+import { isIPad, isDesktop } from "../core/platform/deviceDetection";
 
 // @ts-expect-error Library module
 import { zzfx, zzfxX } from "./sfxPlayer.js";
-// @ts-expect-error Library module
-import CPlayer from "./musicplayer.js";
+import {
+    createTune,
+    FadeIn,
+    FadeOut,
+    FadeOutIn,
+    initMusicPlayer,
+    Tune,
+} from "./music";
 
 export const SFX_START = "start";
 export const SFX_RACE = "race";
@@ -55,28 +61,23 @@ export const SFX_RESTART = "restart";
 export const SFX_COUNT = "count";
 export const SFX_GO = "go";
 
-interface Tune extends HTMLAudioElement {
-    _fadeInterval?: number;
-    _fadeOutInTimeout?: number;
-}
-
-const startTune = document.createElement("audio") as Tune;
-const raceTune = document.createElement("audio") as Tune;
-const gameoverFx = document.createElement("audio") as Tune;
+const startTune = createTune();
+const raceTune = createTune();
+const gameoverFx = createTune();
 
 export let audioUnlocked = false;
 
 // Update the unlockAudio function
-export const unlockAudio = async () => {
+export const unlockAudio = async (): Promise<void> => {
     console.log("Attempting to unlock audio systems...");
 
-    if (audioUnlocked) return true;
+    if (audioUnlocked) return;
 
     // IMPORTANT: Skip unlock only on true desktop browsers, never on iPad
     if (isDesktop && !isIPad) {
         console.log("Skipping audio unlock on desktop browser");
         audioUnlocked = true;
-        return true;
+        return;
     }
 
     console.log("Running audio unlock for mobile/iPad device");
@@ -107,11 +108,10 @@ export const unlockAudio = async () => {
     }
 
     audioUnlocked = true;
-    return true;
 };
 
 // Update the setupAudioUnlock function
-export const setupAudioUnlock = () => {
+export const setupAudioUnlock = (): void => {
     // Skip ONLY on true desktop browsers, never on iPad
     if (isDesktop && !isIPad) {
         console.log("Skipping audio unlock setup on desktop browser");
@@ -171,35 +171,6 @@ export const setupAudioUnlock = () => {
     });
 };
 
-export const initMusicPlayer = (
-    audioTrack: Tune,
-    tune: unknown,
-    isLooped: boolean,
-): Promise<void> => {
-    return new Promise((resolve) => {
-        const songplayer = new CPlayer();
-        // Initialize music generation (player).
-        songplayer.init(tune);
-        // Generate music...
-        let done = false;
-        setInterval(function () {
-            if (done) {
-                return;
-            }
-            done = songplayer.generate() >= 1;
-            if (done) {
-                // Put the generated song in an Audio element.
-                const wave = songplayer.createWave();
-                audioTrack.src = URL.createObjectURL(
-                    new Blob([wave], { type: "audio/wav" }),
-                );
-                audioTrack.loop = isLooped;
-                resolve();
-            }
-        }, 0);
-    });
-};
-
 export const initialize = () => {
     // Set up audio unlock automatically
     setupAudioUnlock();
@@ -209,132 +180,6 @@ export const initialize = () => {
         initMusicPlayer(raceTune, song2, true),
         initMusicPlayer(gameoverFx, gameoverSfx, false),
     ]);
-};
-
-const roundToFractionDigits = (x: number, fractionDigits: number): number =>
-    parseFloat(x.toFixed(fractionDigits));
-
-const FadeOut = (tune: Tune, vol = 0): void => {
-    if (tune._fadeInterval) {
-        clearInterval(tune._fadeInterval);
-        tune._fadeInterval = undefined;
-    }
-    let currentVolume = tune.volume;
-    if (currentVolume > vol) {
-        tune._fadeInterval = setInterval(function () {
-            currentVolume = roundToFractionDigits(
-                Math.max(vol, currentVolume - 0.1),
-                1,
-            );
-            tune.volume = currentVolume;
-            if (currentVolume <= vol) {
-                if (vol === 0) tune.pause();
-                clearInterval(tune._fadeInterval);
-                tune._fadeInterval = undefined;
-            }
-        }, 100);
-    } else if (vol === 0 && !tune.paused) {
-        tune.pause();
-        tune.volume = 0;
-    }
-};
-
-const FadeIn = (tune: Tune, vol: number = 1) => {
-    if (tune._fadeInterval) {
-        clearInterval(tune._fadeInterval);
-        tune._fadeInterval = undefined;
-    }
-
-    let playPromise = Promise.resolve();
-    if (tune.paused) {
-        // Key change: Start with a small audible volume on iOS
-        tune.volume = isIOS ? 0.1 : 0;
-
-        // Add playsinline for iOS (just to be safe)
-        tune.setAttribute("playsinline", "playsinline");
-        playPromise = tune.play();
-    }
-
-    playPromise
-        .then(() => {
-            // Start from current volume
-            let currentVolume = tune.volume;
-
-            if (currentVolume < vol) {
-                if (tune._fadeInterval) clearInterval(tune._fadeInterval);
-
-                tune._fadeInterval = setInterval(function () {
-                    currentVolume = roundToFractionDigits(
-                        Math.min(vol, currentVolume + 0.1),
-                        1,
-                    );
-                    tune.volume = currentVolume;
-
-                    if (currentVolume >= vol) {
-                        tune.volume = vol;
-                        clearInterval(tune._fadeInterval);
-                        tune._fadeInterval = undefined;
-                    }
-                }, 100);
-            } else {
-                tune.volume = vol;
-            }
-        })
-        .catch((e) => {
-            console.warn("FadeIn play() failed:", e);
-            // Try one more time with higher volume on any failure
-            if (tune.paused) {
-                tune.volume = 1;
-                tune.play().catch((err) =>
-                    console.error("Second play attempt failed:", err),
-                );
-            }
-
-            if (tune._fadeInterval) {
-                clearInterval(tune._fadeInterval);
-                tune._fadeInterval = undefined;
-            }
-        });
-};
-
-const FadeOutIn = (tune1: Tune, tune2: Tune): void => {
-    if (tune1._fadeInterval) clearInterval(tune1._fadeInterval);
-    if (tune1._fadeOutInTimeout) clearTimeout(tune1._fadeOutInTimeout);
-    if (tune2._fadeInterval) clearInterval(tune2._fadeInterval);
-    if (tune2._fadeOutInTimeout) clearTimeout(tune2._fadeOutInTimeout);
-    tune1._fadeInterval = undefined;
-    tune1._fadeOutInTimeout = undefined;
-    tune2._fadeInterval = undefined;
-    tune2._fadeOutInTimeout = undefined;
-
-    let currentVolume = tune1.volume;
-    if (currentVolume > 0) {
-        tune1._fadeInterval = setInterval(function () {
-            currentVolume = roundToFractionDigits(
-                Math.max(0, currentVolume - 0.1),
-                1,
-            );
-            tune1.volume = currentVolume;
-
-            if (currentVolume <= 0.1) {
-                tune1.pause();
-                clearInterval(tune1._fadeInterval);
-                tune1._fadeInterval = undefined;
-
-                tune1._fadeOutInTimeout = setTimeout(() => {
-                    FadeIn(tune2, 1);
-                    tune1._fadeOutInTimeout = undefined;
-                }, 500);
-            }
-        }, 100);
-    } else {
-        tune1.pause();
-        tune1.volume = 0;
-        tune1._fadeOutInTimeout = setTimeout(() => {
-            FadeIn(tune2, 1);
-            tune1._fadeOutInTimeout = undefined;
-        }, 500);
-    }
 };
 
 export const playTune = async (
@@ -402,7 +247,7 @@ export const playTune = async (
     }
 };
 
-export const stopTune = (tune?: string) => {
+export const stopTune = (tune?: string): void => {
     const tunesToStop: Tune[] = [];
 
     if (tune === SFX_RACE) {
